+++ backend/src/main/java/hanglog/trip/domain/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.trip.domain.category.repository;
 
-import hanglog.trip.domain.Category;
+import hanglog.trip.domain.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
