+++ backend/src/main/java/hanglog/trip/domain/city/City.java
@@ -1,4 +1,4 @@
-package hanglog.trip.domain;
+package hanglog.trip.domain.city;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
