+++ backend/src/main/java/hanglog/trip/domain/daylog/repository/DayLogRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.trip.domain.daylog.repository;
 
-import hanglog.trip.domain.DayLog;
+import hanglog.trip.domain.daylog.DayLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DayLogRepository extends JpaRepository<DayLog, Long> {
