+++ backend/src/main/java/hanglog/trip/domain/expense/Expense.java
@@ -1,40 +1,52 @@
-package hanglog.trip.domain;
+package hanglog.trip.domain.expense;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import jakarta.persistence.Basic;
+import hanglog.trip.domain.item.Item;
+import hanglog.trip.domain.category.Category;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class Picture {
+public class Expense {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    @Lob
-    @Basic(fetch = LAZY)
     @Column(nullable = false)
-    private byte[] blobImg;
+    private String currency;
 
-    @ManyToOne(fetch = LAZY)
+    @Column(nullable = false)
+    private Integer amount;
+
+    @OneToOne(fetch = LAZY)
     @JoinColumn(name = "item_id", nullable = false)
     private Item item;
 
-    public Picture(final byte[] blobImg, final Item item) {
-        this.blobImg = blobImg;
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "category_id", nullable = false)
+    private Category category;
+
+    public Expense(String currency,
+                   Integer amount,
+                   Item item,
+                   Category category
+    ) {
+        this.currency = currency;
+        this.amount = amount;
         this.item = item;
+        this.category = category;
     }
 }
