+++ backend/src/main/java/hanglog/trip/domain/expense/repository/ExpenseRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.trip.domain.expense.repository;
 
-import hanglog.trip.domain.Expense;
+import hanglog.trip.domain.expense.Expense;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ExpenseRepository extends JpaRepository<Expense, Long> {
