+++ backend/src/main/java/hanglog/trip/domain/picture/Picture.java
@@ -1,14 +1,17 @@
-package hanglog.trip.domain;
+package hanglog.trip.domain.picture;
 
 import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
-import hanglog.global.BaseTimeEntity;
+import hanglog.trip.domain.item.Item;
+import jakarta.persistence.Basic;
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,22 +19,23 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-public class TripCity extends BaseTimeEntity {
+public class Picture {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
-    @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "trip_id")
-    private Trip trip;
+    @Lob
+    @Basic(fetch = LAZY)
+    @Column(nullable = false)
+    private byte[] blobImg;
 
     @ManyToOne(fetch = LAZY)
-    @JoinColumn(name = "city_id")
-    private City city;
+    @JoinColumn(name = "item_id", nullable = false)
+    private Item item;
 
-    public TripCity(final Trip trip, final City city) {
-        this.trip = trip;
-        this.city = city;
+    public Picture(final byte[] blobImg, final Item item) {
+        this.blobImg = blobImg;
+        this.item = item;
     }
 }
