+++ backend/src/main/java/hanglog/trip/domain/place/repository/PlaceRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.trip.domain.place.repository;
 
-import hanglog.trip.domain.Place;
+import hanglog.trip.domain.place.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceRepository extends JpaRepository<Place, Long> {
