+++ backend/src/main/java/hanglog/trip/domain/trip/repository/TripRepository.java
@@ -1,6 +1,6 @@
-package hanglog.trip.domain.repository;
+package hanglog.trip.domain.trip.repository;
 
-import hanglog.trip.domain.Trip;
+import hanglog.trip.domain.trip.Trip;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TripRepository extends JpaRepository<Trip, Long> {
